Extract interfaces and move to sub-namespaces
